rebase master

+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/MaskHeadSpec.scala
@@ -338,7 +338,7 @@ class MaskHeadSpec extends FlatSpec with Matchers {
     val inChannels: Int = 6
     val resolution: Int = 14
     val scales: Array[Float] = Array[Float](0.25f, 0.125f)
-    val samplingRratio: Float = 2.0f
+    val samplingRratio: Int = 2
     val layers: Array[Int] = Array[Int](4, 4)
     val dilation: Int = 1
     val numClasses: Int = 81
